+++ app/router/web.php
@@ -23,11 +23,11 @@ Router::add("GET","/signIn","ClientViewController@create");
 Router::add("POST","/store","ClientController@store");
 Router::add("POST","approuver/{id}","CompteController@approuver");
 Router::add("POST","/refuser/{id}","CompteController@refuser");
-Router::add("GET","/demandeCompte","ClientViewController@demandeComptes");
-Router::add("GET","/voir/{id}","ClientViewController@demandeCompte");
+Router::add("GET","/demandeCompte","EmployeViewController@demandeComptes");
+Router::add("GET","/voir/{id}","EmployeViewController@demandeCompte");
 Router::add("POST","/addClient","ClientController@add");
-Router::add("GET","/clients","ClientViewController@clients");
-Router::add("POST","/client/{id}","ClientViewController@client");
+Router::add("GET","/clients","EmployeViewController@clients");
+Router::add("POST","/client/{id}","EmployeViewController@client");
 Router::add("POST","client/delete/{id}","ClientController@delete");
 Router::add("GET","/Client","ClientViewController@index");
 Router::add("GET","update-info","ClientViewController@edite");
